+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 ### Fixed
 - rename results' index from 'raster_id_100m' to 'index' to support the grid value visualization in dashboard([41](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/41))
 
+### Changed
+- Introduced artifact tags ([40](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/40))
+
 ### Added
 - Maps of dominant building ages and energy carriers ([28](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/28))